discord-plugin/pyproject.toml
[tool.poetry]
name = "voidling_discord"
version = "0.1.0"
description = ""
authors = ["Aditya Rathod <15635979+adityarathod@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
pydantic = "^1.10.7"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
